reafactor: Moved SerieSearchOption to Common
