+++ src/main/java/edu/lorsenmarek/backend/common/SerieSearchOption.java
@@ -1,4 +1,4 @@
-package edu.lorsenmarek.backend.utility;
+package edu.lorsenmarek.backend.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
